MARL-curve-main/requirements_taoan.txt
absl-py==1.2.0
aniso8601==9.0.1
ansi2html==1.8.0
anyio==3.6.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.0.8
atomicwrites==1.2.1
attrs==18.2.0
awscli==1.22.101
Babel==2.10.3
backcall==0.2.0
backports.functools-lru-cache==1.6.4
beautifulsoup4==4.11.1
bleach==5.0.1
blis==0.7.8
bokeh==2.4.3
boto3==1.21.46
botocore==1.24.46
brotlipy==0.7.0
captum==0.5.0
catalogue==2.0.8
certifi==2022.6.15
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.0
colorama==0.4.5
cryptography==37.0.4
cycler==0.10.0
cymem==2.0.6
dataclasses==0.8
debugpy==1.6.3
decorator==5.1.1
deepdiff==5.8.1
defusedxml==0.7.1
dill==0.3.5.1
dm-env==1.5
dm-env-rpc==1.1.3
dm-tree==0.1.7
docopt==0.6.2
docutils==0.15.2
dparse==0.5.2
entrypoints==0.4
enum34==1.1.10
executing==1.0.0
fastai==2.1.10
fastcore==1.5.25
fastjsonschema==2.16.1
fastprogress==1.0.3
filelock==3.6.0
Flask==2.2.2
Flask-RESTful==0.3.9
flit_core==3.7.1
fonttools==4.37.1
fsspec==2022.8.2
future==0.16.0
gitdb==4.0.9
GitPython==3.1.27
google-pasta==0.2.0
googleapis-common-protos==1.56.4
grpcio==1.47.0
gym==0.23.1
gym-notices==0.0.8
horovod==0.25.0
idna==3.3
imageio==2.4.1
immutabledict==2.2.1
importlib-metadata==4.11.4
importlib-resources==5.9.0
ipykernel==6.15.2
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.1.0
json5==0.9.5
jsonpickle==1.5.2
jsonschema==4.15.0
jupyter_client==7.3.5
jupyter_core==4.11.1
jupyter-server==1.18.1
jupyterlab==3.3.4
jupyterlab-pygments==0.2.2
jupyterlab_server==2.15.1
jupyterlab-widgets==3.0.3
kiwisolver==1.0.1
langcodes==3.3.0
llvmlite==0.39.1
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.0.0
matplotlib-inline==0.1.6
mistune==2.0.4
mock==4.0.3
more-itertools==4.3.0
mpyq==0.2.5
munch==2.5.0
munkres==1.1.4
murmurhash==1.0.8
nbclassic==0.4.3
nbclient==0.6.7
nbconvert==7.0.0
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.12
notebook-shim==0.1.0
numba==0.56.2
numpy==1.23.2
nvgpu==0.9.0
opencv-python==4.6.0.66
ordered-set==4.1.0
packaging==21.3
pandas==1.4.4
pandocfilters==1.5.0
parso==0.8.3
pathlib2==2.3.2
pathy==0.6.2
patsy==0.5.2
pbr==4.3.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pip==22.2.2
pipreqs==0.4.11
pkgutil_resolve_name==1.3.10
plotly==5.6.0
pluggy==0.7.1
ply==3.11
portpicker==1.5.2
preshed==3.0.7
probscale==0.2.3
prometheus-client==0.14.1
prompt-toolkit==3.0.31
protobuf==3.20.1
protobuf3-to-dict==0.1.5
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.6.0
py-cpuinfo==8.0.0
pyarrow==9.0.0
pyasn1==0.4.8
pybind11==2.9.2
pybind11-global==2.9.2
pycparser==2.21
pydantic==1.9.2
pyfunctional==1.4.3
pygame==2.1.2
Pygments==2.13.0
pynvml==11.4.1
pyOpenSSL==22.0.0
pyparsing==3.0.9
PyQt5==5.15.7
PyQt5-sip==12.11.0
pyrsistent==0.18.1
PySC2==4.0.0
PySocks==1.7.1
pytest==3.8.2
python-dateutil==2.7.3
pytz==2022.2.1
PyYAML==3.13
pyzmq==23.2.1
requests==2.28.1
rsa==4.7.2
s2clientprotocol==5.0.10.88500.0
s2protocol==5.0.10.88500.0
s3fs==0.4.2
s3transfer==0.5.2
sacred==0.8.2
safety==1.10.3
sagemaker==2.22.0
scikit-learn==1.0
scipy==1.9.1
seaborn==0.11.2
Send2Trash==1.8.0
setuptools==65.3.0
shap==0.40.0
shellingham==1.5.0
sip==6.6.2
six==1.16.0
sk-video==1.1.10
sklearn==0.0
slicer==0.0.7
SMAC==1.0.0
smart-open==5.2.1
smclarify==0.2
smdebug-rulesconfig==1.0.0
smmap==5.0.0
snakeviz==1.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
spacy==3.4.1
spacy-legacy==3.0.10
spacy-loggers==1.0.3
srsly==2.4.4
stack-data==0.5.0
statsmodels==0.13.2
tabulate==0.8.10
tenacity==8.0.1
tensorboard-logger==0.1.0
termcolor==1.1.0
terminado==0.15.0
thinc==8.1.0
threadpoolctl==3.1.0
tinycss2==1.1.1
toml==0.10.2
torch==1.12.1
torch-model-archiver==0.5.3b20220226
torch-workflow-archiver==0.2.4b20220513
torchaudio==0.12.1
torchserve==0.6.0b20220513
torchtext==0.13.0
torchvision==0.2.2
tornado==6.2
tqdm==4.19.9
traitlets==5.3.0
typer==0.4.2
typing_extensions==4.3.0
unicodedata2==14.0.0
urllib3==1.26.12
wasabi==0.10.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.0
Werkzeug==2.2.2
wheel==0.37.1
whichcraft==0.5.2
widgetsnbextension==4.0.3
wrapt==1.14.1
yarg==0.1.9
zipp==3.8.1
git+https://github.com/oxwhirl/smac.git
